remove the mapping of <Esc> on normal mode
the mapping causes a some kind of error on startup time.
(In my case, Vim is starting with Operator-Pending mode of s command.)

On the startup, Vim sends some commands which contains <Esc> to get
terminal info. If the mapping is defined, it eats <Esc> of the commands
and causes the unexpected behaviors.